remove UpdateHandler
The new version of the temporal SDK changes the signature of the
UpdateWorkflow method. To keep compatibility with old and new versions
of the SDK, I'm removing the method from this library. If anyone wants
it back, I can add it with the new SDK version in a future release.