remove namespace safety features
It turns out that with temporal cloud, different instances of a service
need to connect to different namespace. Ex. one for staging and one for
production. The way that the namespace safety features were designed,
they required a static namespace name to be declared for each queue.
This made it basically impossible to use tempts with temporal cloud. I'm
open to bringing back the namespace safety features once there's a
solution for using dynamic namespace names. I'm trying to avoid using
global state to remap namespace names. If anyone has a better idea, let
me know.